start deprecating old cross_section
